requirements.txt
# Auto-generated from pyproject.toml - do not edit manually
# Run: make sync-requirements

# Core web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
aiosqlite>=0.19.0

# Task queue (optional for HA)
celery[redis]>=5.4.0
redis>=5.2.0

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0
apscheduler>=3.10.0

# MCP Server
fastmcp>=2.0.0

# Crypto & Finance
yfinance>=0.2.0
statsforecast>=1.7.0
numpy>=1.26.0
pandas>=2.0.0

# HTTP & Network
websockets>=12.0

# Other dependencies
greenlet>=3.3.0
chronos-forecasting>=1.4.0
torch>=2.0.0
neuralprophet>=0.8.0
optuna>=3.5.0
